back/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
SQLAlchemy>=2.0,<3.0
aiosqlite>=0.21,<1.0
alembic>=1.15,<2.0
pydantic>=2.10,<3.0
pydantic-settings>=2.8,<3.0
greenlet>=3.1,<4.0
google-genai>=1.0,<2.0
pyyaml>=6.0,<7.0
